simplify the behavior - stop exposing it as a subcommand
